requirements.txt
psycopg2-binary==2.8.3
bokeh==1.2.0
